backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
scikit-learn==1.4.0
pandas==2.2.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
bcrypt==4.1.2
email-validator==2.1.1
httpx==0.25.2
